Add Postgres Dockerfile and db setup instructions
Introduced a Dockerfile for running PostgreSQL in a Docker container and also added a readme file that includes instructions for setting up PostgreSQL, installing and running it on local machine using Docker. The instructions also cover establishing a successful connection and ensuring that server credentials are properly specified to avoid errors.