Backend modular (#3)
* make root package.json

* Add start:backend script to package.json

A new script to start the backend service has been added to package.json. This provides an easy, standardized way to start the backend server for the mq-chat application.

* Add package.json and package-lock.json with essential back-end dependencies

This commit initiates the package.json and package-lock.json files, installing necessary dependencies such as express, graphql, and kafkajs. It also includes '@types/express' as a devDependency to support Typescript in the back-end.

* Create basic express server with GraphQL setup

This commit introduces a basic express server setup with the initial GraphQL schema and resolver. It also includes placeholder code for schema construction and api endpoint function definitions. The server is meant to run on port 4000.

* Add Socket.io to server and update dependencies

Socket.io server support has been implemented, creating a standalone Express server parsed into an Express app that's wrapped within a Socket.io server. Also included necessary imports and instances of Socket.io and the HTTP server wrapper. To maintain optimal performance, Socket.io has been added to dependencies on the backend.

* Add Postgres Dockerfile and db setup instructions

Introduced a Dockerfile for running PostgreSQL in a Docker container and also added a readme file that includes instructions for setting up PostgreSQL, installing and running it on local machine using Docker. The instructions also cover establishing a successful connection and ensuring that server credentials are properly specified to avoid errors.

* Add Postgres Dockerfile and db setup instructions

Introduced a Dockerfile for running PostgreSQL in a Docker container and also added a readme file that includes instructions for setting up PostgreSQL, installing and running it on local machine using Docker. The instructions also cover establishing a successful connection and ensuring that server credentials are properly specified to avoid errors.

* Shelving postgraphile due to reaching a dead-end

Need to read into postgres schemas and how to define them properly, also monolith backend is difficult to implement with both postgraphile and socket.io

* rebuild project around backend microservices