Add new data sources and setup GraphQL module
Three new XML data sources have been added to configure the MongoDB database connection used in the project. The XML files contain specification for the MongoDB database, setup details, connection URLs and other DB information. In addition, a new GraphQL module is set up in the frontend which uses ApolloClientOptions to establish and define the link between the backend and the Angular frontend. The URI of the GraphQL server is configured in this module. This GraphQL setup allows queries to the backend server, providing data management functionalities.